+++ example/src/CrudWebserverFile.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings     #-}
 {-# LANGUAGE StandaloneDeriving    #-}
+{-# LANGUAGE TemplateHaskell       #-}
 {-# LANGUAGE TypeOperators         #-}
 
 ------------------------------------------------------------------------
@@ -73,6 +74,8 @@ import           Test.QuickCheck.Instances
                    ()
 
 import           Test.StateMachine
+import           Test.StateMachine.TH
+                   (deriveTestClasses)
 
 ------------------------------------------------------------------------
 
@@ -203,23 +206,10 @@ semantics act = do
 
 ------------------------------------------------------------------------
 
-instance HTraversable Action where
-  htraverse _ (PutFile    file content) = pure (PutFile    file content)
-  htraverse _ (GetFile    file)         = pure (GetFile    file)
-  htraverse _ (DeleteFile file)         = pure (DeleteFile file)
-
-instance HFunctor  Action
-instance HFoldable Action
+deriveTestClasses ''Action
 
 instance Show (Untyped Action) where
   show (Untyped act) = show act
-
-instance Constructors Action where
-  constructor x = Constructor $ case x of
-    PutFile{}    -> "PutFile"
-    GetFile{}    -> "GetFile"
-    DeleteFile{} -> "DeleteFile"
-  nConstructors _ = 3
 
 ------------------------------------------------------------------------