+++ src/Test/StateMachine.hs
@@ -48,7 +48,6 @@ module Test.StateMachine
   , monadicParallelC
 
     -- * Types
-  , module Test.StateMachine.Generics
   , module Test.StateMachine.Types
 
     -- * Reexport
@@ -75,7 +74,6 @@ import qualified Test.QuickCheck.Counterexamples       as CE
 import           Test.QuickCheck.Monadic
                    (PropertyM, monadic, run)
 
-import           Test.StateMachine.Generics
 import           Test.StateMachine.Internal.Parallel
 import           Test.StateMachine.Internal.Sequential
 import           Test.StateMachine.Internal.Types